Refactor helpers.
