+++ lib/http.js
@@ -6,7 +6,6 @@
 const {asyncHandler} = require('bedrock-express');
 const bedrock = require('bedrock');
 const {
-  defaultDocumentLoader,
   runOperation,
   validateOperation,
   keystores,
@@ -17,17 +16,10 @@ const {createMiddleware} = require('bedrock-passport');
 const brZCapStorage = require('bedrock-zcap-storage');
 const cors = require('cors');
 const database = require('bedrock-mongodb');
-const {verifyCapabilityInvocation} = require('http-signature-zcap-verify');
-const jsigs = require('jsonld-signatures');
-const {extendContextLoader, SECURITY_CONTEXT_V2_URL} = jsigs;
-const {Ed25519Signature2018} = jsigs.suites;
-const {CapabilityDelegation} = require('ocapld');
 const URL = require('url').URL;
 const {generateRandom} = require('webkms-switch');
+const helpers = require('./helpers');
 require('bedrock-express');
-
-// load config defaults
-require('./config');
 
 bedrock.events.on('bedrock-express.configure.routes', app => {
   const cfg = config['kms-http'];
@@ -39,13 +31,6 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
   routes.keys = `${routes.keystore}/keys`;
   routes.key = `${routes.keys}/:keyId`;
   routes.zcaps = `${routes.keystore}/zcaps`;
-
-  function _getKeystoreId({host, req, localId}) {
-    if(!host) {
-      host = req.get('host');
-    }
-    return `https://${host}${routes.keystores}/${localId}`;
-  }
 
   // TODO: endpoints for creating and deleting keystores will only use
   // session-based auth and check that an account exists on the system
@@ -81,7 +66,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       }
 
       const random = await generateRandom();
-      const id = _getKeystoreId({req, localId: random});
+      const id = helpers.getKeystoreId({req, localId: random, routes});
       _verifyHost(id);
 
       // create a keystore for the controller
@@ -152,7 +137,8 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       // changes to make and then post them to the non-proxied update
       // config endpoint using `sequence` to ensure atomicity of transaction
 
-      const id = _getKeystoreId({req, localId: req.params.keystoreId});
+      const id = helpers.getKeystoreId(
+        {req, localId: req.params.keystoreId, routes});
       _verifyHost(id);
 
       const {config} = await keystores.get({id});
@@ -195,7 +181,8 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     // TODO: consider making this zcap authorized instead
     createMiddleware({strategy: 'session'}),
     asyncHandler(async (req, res) => {
-      const id = _getKeystoreId({req, localId: req.params.keystoreId});
+      const id = helpers.getKeystoreId(
+        {req, localId: req.params.keystoreId, routes});
       _verifyHost(id);
       const {config} = await keystores.get({id});
       res.json(config);
@@ -221,7 +208,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       const host = req.get('host');
       const id = `https://${host}${req.originalUrl}`;
       _verifyHost(id);
-      const result = _getInvocationTarget({host, url: id});
+      const result = helpers.getInvocationTarget({host, url: id});
       if(!result) {
         // invalid root zcap ID
         throw new BedrockError(
@@ -232,7 +219,8 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       const {target, keystoreId} = result;
 
       // dynamically generate root capability for target
-      const zcap = await _generateRootCapability({id, target, keystoreId});
+      const zcap = await helpers.generateRootCapability(
+        {id, target, keystoreId});
       res.json(zcap);
     }));
 
@@ -244,7 +232,8 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     cors(),
     asyncHandler(async (req, res) => {
       // expect root capability to be `.../zcaps/keys`
-      const keystoreId = _getKeystoreId({req, localId: req.params.keystoreId});
+      const keystoreId = helpers.getKeystoreId(
+        {req, localId: req.params.keystoreId, routes});
       const expectedRootCapability = `${keystoreId}/zcaps/keys`;
       const result = await _handleOperation({
         req, expectedRootCapability, keystoreId
@@ -267,7 +256,8 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     cors(),
     asyncHandler(async (req, res) => {
       // dynamically generate zcap for root capability
-      const keystoreId = _getKeystoreId({req, localId: req.params.keystoreId});
+      const keystoreId = helpers.getKeystoreId(
+        {req, localId: req.params.keystoreId, routes});
       const id = `${keystoreId}/keys/${req.params.keyId}`;
       const {key} = await keyDescriptionStorage.get({id});
       res.json(key);
@@ -284,11 +274,11 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     //validate('bedrock-kms-http.zcap'),
     asyncHandler(async (req, res) => {
       // check authorization
-      const keystoreId = _getKeystoreId(
-        {req, localId: req.params.keystoreId});
+      const keystoreId = helpers.getKeystoreId(
+        {req, localId: req.params.keystoreId, routes});
       const expectedTarget = `${keystoreId}/authorizations`;
       const expectedRootCapability = `${keystoreId}/zcaps/authorizations`;
-      const {invoker} = await _authorize({
+      const {invoker} = await helpers.authorize({
         req, expectedTarget, expectedRootCapability,
         expectedAction: 'write'
       });
@@ -297,7 +287,8 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       const controller = invoker;
       const capability = req.body;
       const host = req.get('host');
-      await _verifyDelegation({keystoreId, host, controller, capability});
+      await helpers.verifyDelegation(
+        {keystoreId, host, controller, capability});
       await brZCapStorage.authorizations.insert({controller, capability});
       res.status(204).end();
     }));
@@ -307,11 +298,11 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     routes.authorizations,
     asyncHandler(async (req, res) => {
       // check authorization
-      const keystoreId = _getKeystoreId(
-        {req, localId: req.params.keystoreId});
+      const keystoreId = helpers.getKeystoreId(
+        {req, localId: req.params.keystoreId, routes});
       const expectedTarget = `${keystoreId}/authorizations`;
       const expectedRootCapability = `${keystoreId}/zcaps/authorizations`;
-      const {invoker} = await _authorize({
+      const {invoker} = await helpers.authorize({
         req, expectedTarget, expectedRootCapability,
         expectedAction: 'read'
       });
@@ -338,11 +329,11 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     cors(),
     asyncHandler(async (req, res) => {
       // check authorization
-      const keystoreId = _getKeystoreId(
-        {req, localId: req.params.keystoreId});
+      const keystoreId = helpers.getKeystoreId(
+        {req, localId: req.params.keystoreId, routes});
       const expectedTarget = `${keystoreId}/authorizations`;
       const expectedRootCapability = `${keystoreId}/zcaps/authorizations`;
-      const {invoker} = await _authorize({
+      const {invoker} = await helpers.authorize({
         req, expectedTarget, expectedRootCapability,
         expectedAction: 'write'
       });
@@ -383,8 +374,8 @@ async function _handleOperation({req, expectedRootCapability, keystoreId}) {
     operation,
     expectedHost: config.kms.allowedHost,
     expectedRootCapability,
-    getInvokedCapability: createGetInvokedCapability(host),
-    documentLoader: createCapabilityLoader({host, expectedTarget: url})
+    getInvokedCapability: helpers.createGetInvokedCapability({host}),
+    documentLoader: helpers.createCapabilityLoader({host, expectedTarget: url})
   });
   if(!result.valid) {
     if(result.error instanceof SyntaxError) {
@@ -421,248 +412,6 @@ async function _handleOperation({req, expectedRootCapability, keystoreId}) {
 
 // TODO: a lot of helper code here is in common w/ bedrock-data-hub-storage,
 // consider refactoring to share it
-
-async function _authorize({
-  req, expectedTarget, expectedRootCapability, expectedAction
-}) {
-  const host = req.get('host');
-  const url = `https://${host}${req.originalUrl}`;
-  const {method, headers} = req;
-  const result = await verifyCapabilityInvocation({
-    url, method, headers,
-    getInvokedCapability: createGetInvokedCapability(host),
-    documentLoader: createCapabilityLoader({host, expectedTarget}),
-    expectedHost: config.server.host,
-    expectedTarget, expectedRootCapability, expectedAction,
-    // TODO: support RsaSignature2018 and other suites?
-    suite: [new Ed25519Signature2018()]
-  });
-  if(!result.verified) {
-    throw new BedrockError(
-      'Permission denied.', 'NotAllowedError', {
-        httpStatusCode: 400,
-        public: true
-      }, result.error);
-  }
-  return {
-    valid: result.verified,
-    ...result
-  };
-}
-
-/**
- * Gets an `expectedRootCapability` from an `invocationTarget`. This function
- * handles special cases where the `invocationTarget` does not match the
- * `expectedRootCapability` because the `invocationTarget` cannot be expressed
- * as a zcap.
- *
- * @param {object} options - Options to use.
- * @param {string} options.invocationTarget - An invocationTarget.
- * @param {string} options.keystoreId - A url to a key store.
- *
- * @return {string} The `expectedRootCapability` for the `invocationTarget`.
- */
-function _getExpectedRootCapability({invocationTarget, keystoreId}) {
-  // `authorizations` endpoint cannot be expressed as a zcap, map to zcap space
-  if(invocationTarget === `${keystoreId}/authorizations`) {
-    return `${keystoreId}/zcaps/authorizations`;
-  }
-  // `keys` endpoint cannot be expressed as a zcap, map to zcap space
-  if(invocationTarget === `${keystoreId}/keys`) {
-    return `${keystoreId}/zcaps/keys`;
-  }
-  // otherwise use default behavior of `invocationTarget` matches root zcap
-  return invocationTarget;
-}
-
-async function _verifyDelegation({
-  keystoreId, host, /*controller,*/capability
-}) {
-  const invocationTarget = typeof capability.invocationTarget === 'string' ?
-    capability.invocationTarget : capability.invocationTarget.id;
-  const documentLoader = extendContextLoader(async url => {
-    // if `id` starts with `<keystoreId>/keys/`, assume zcap is a key
-    if(url.startsWith(`${keystoreId}/keys/`)) {
-      // dynamically generate zcap for root capability
-      const {key} = await keyDescriptionStorage.get({id: url});
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: key
-      };
-    }
-
-    // check if URL is a root zcap w/ a different invocation target
-    const result = _getInvocationTarget({host, url});
-    if(result) {
-      // dynamically generate zcap for root capability
-      const {target, keystoreId} = result;
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: await _generateRootCapability({id: url, target, keystoreId})
-      };
-    }
-
-    // see if zcap is in storage
-    try {
-      const {authorization} = await brZCapStorage.authorizations.get(
-        {id: url, invocationTarget});
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: authorization.capability
-      };
-    } catch(e) {
-      if(e.name !== 'NotFoundError') {
-        throw e;
-      }
-    }
-
-    return defaultDocumentLoader(url);
-  });
-  const expectedRootCapability = _getExpectedRootCapability(
-    {invocationTarget, keystoreId});
-  const {verified, error} = await jsigs.verify(capability, {
-    suite: new Ed25519Signature2018(),
-    purpose: new CapabilityDelegation({
-      expectedRootCapability,
-      suite: new Ed25519Signature2018()
-    }),
-    documentLoader,
-    compactProof: false
-  });
-  if(!verified) {
-    throw error;
-  }
-}
-
-// wrap document loader to always generate root zcap from keystore config
-// description in storage
-function createCapabilityLoader({host, expectedTarget}) {
-  return async function rootCapabilityLoader(url) {
-    const result = _getInvocationTarget({host, url});
-    if(result) {
-      // dynamically generate zcap for root capability
-      const {target, keystoreId} = result;
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: await _generateRootCapability({id: url, target, keystoreId})
-      };
-    }
-
-    // see if zcap is in storage
-    try {
-      const {authorization} = await brZCapStorage.authorizations.get(
-        {id: url, invocationTarget: expectedTarget});
-      return {
-        contextUrl: null,
-        documentUrl: url,
-        document: authorization.capability
-      };
-    } catch(e) {
-      if(e.name !== 'NotFoundError') {
-        throw e;
-      }
-    }
-
-    return defaultDocumentLoader(url);
-  };
-}
-
-function createGetInvokedCapability(host) {
-  return async function getInvokedCapability({id, expectedTarget}) {
-    try {
-      // if `id` matches `expectedTarget`, assume zcap is a key
-      if(id === expectedTarget) {
-        // dynamically generate zcap for root capability
-        const {key} = await keyDescriptionStorage.get({id});
-        return key;
-      }
-
-      // if the capability is a root zcap generated by this server then its
-      // `id` will map to an invocation target; if so, dynamically generate the
-      // zcap as it is the root authority which is automatically authorized
-      const result = _getInvocationTarget({host, url: id});
-      if(result) {
-        // dynamically generate zcap for root capability
-        const {target, keystoreId} = result;
-        return _generateRootCapability({id, target, keystoreId});
-      }
-
-      // otherwise, must get capability from authorizations storage
-      const {authorization} = await brZCapStorage.authorizations.get({
-        id,
-        invocationTarget: expectedTarget
-      });
-      return authorization.capability;
-    } catch(e) {
-      if(e.name === 'NotFoundError') {
-        throw new BedrockError(
-          'Permission denied.', 'NotAllowedError', {
-            httpStatusCode: 400,
-            public: true
-          }, e);
-      }
-      throw e;
-    }
-  };
-}
-
-async function _generateRootCapability(
-  {id, target, keystoreId, config = null}) {
-  if(!config) {
-    ({config} = await keystores.get({id: keystoreId}));
-  }
-  // dynamically generate zcap for root capability
-  return {
-    '@context': SECURITY_CONTEXT_V2_URL,
-    id,
-    invocationTarget: target,
-    controller: config.controller,
-    invoker: config.invoker,
-    delegator: config.delegator
-  };
-}
-
-function _getInvocationTarget({host, url}) {
-  // look for `/kms/keystores/<keystoreId>/zcaps/`
-  const cfg = config['kms-http'];
-  const baseKeystoreUrl = `https://${host}${cfg.routes.basePath}/keystores/`;
-  let idx = url.indexOf(baseKeystoreUrl);
-  if(idx !== 0) {
-    return null;
-  }
-
-  // skip keystore ID
-  const keystoreIdIdx = baseKeystoreUrl.length;
-  idx = url.indexOf('/', keystoreIdIdx);
-  if(idx === -1) {
-    return null;
-  }
-  const keystoreId = `${baseKeystoreUrl}${url.substring(keystoreIdIdx, idx)}`;
-
-  // skip `zcaps`
-  idx = url.indexOf('zcaps/', idx + 1);
-  if(idx === -1) {
-    return null;
-  }
-
-  // valid root zcap invocation targets:
-  // `/kms/keystores/<keystoreId>/keys`
-  // `/kms/keystores/<keystoreId>/authorizations`
-  const path = url.substr(idx + 6 /* 'zcaps/'.length */);
-  if(!['keys', 'authorizations'].includes(path)) {
-    return null;
-  }
-
-  // return invocation target for the given root zcap URL
-  return {
-    target: `${keystoreId}/${path}`,
-    keystoreId
-  };
-}
 
 function _verifyHost(url) {
   const {allowedHost} = config.kms;
