+++ lib/index.js
@@ -1,6 +1,10 @@
 /*!
- * Copyright (c) 2019 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2020 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
 require('./http');
+require('./http-revocations');
+
+// load config defaults
+require('./config');
